fix domain module nucleic entity filed
